Fixing all ios8 warnings. v7.0.0 is ready to go to TestFlight!

+++ clients/ios/Classes/NewsBlurViewController.h
@@ -122,8 +122,8 @@ UIGestureRecognizerDelegate> {
 - (void)showRefreshNotifier;
 - (void)showCountingNotifier;
 - (void)showSyncingNotifier;
-- (void)showSyncingNotifier:(float)progress hoursBack:(int)days;
-- (void)showCachingNotifier:(float)progress hoursBack:(int)hours;
+- (void)showSyncingNotifier:(float)progress hoursBack:(NSInteger)days;
+- (void)showCachingNotifier:(float)progress hoursBack:(NSInteger)hours;
 - (void)showOfflineNotifier;
 - (void)showDoneNotifier;
 - (void)hideNotifier;
